logo x6
Garett and Anatol helped / did all the hard work to get a logo to the
screen. We even got 6 logos to tile the screen (though more than that
will be difficult with the current approach)

We did this by writing a Ruby script that emits 6502 asm code that
writes the correct bytes to the correct locations in memory. (Cooool.)